requirements.txt
numpy==1.26.4
matplotlib==3.9.2
scipy==1.14.1
sounddevice==0.5.1
ttkthemes==3.2.2
pycryptodome==3.21.0
